requirements.txt

Django==3.2
django-crispy-forms==1.11.2
django-markupfield==2.0.0
django-recaptcha==2.0.6
django-versatileimagefield==2.0
Markdown==3.3.4
Pillow==8.2.0
